MWE: Rust on Nucleo L432KCU board
memory.x is a symlink to what will become a collection of memory layouts
based on datasheets of individual chips. Check with `cargo check`; build
with `cargo build`; deploy with `cargo run`, which uses probe-rs, which
just needs the name of the chip and permission to read-write it.
onfigure memory.x, .cargo/config.toml, and Cargo.toml to reflect chip
selection.